some fixes

+++ accaunts/models.py
@@ -316,8 +316,8 @@ class UserBet(models.Model):
         (HEART, 'червы'),
         (SPADES, 'пики'),
     ]
-    sum = models.PositiveIntegerField(verbose_name='Сумма ставки', default=0)
-    sum_win = models.PositiveIntegerField(verbose_name='Сумма выигрыша', default=0)
+    sum = models.PositiveBigIntegerField(verbose_name='Сумма ставки', default=0)
+    sum_win = models.PositiveBigIntegerField(verbose_name='Сумма выигрыша', default=0)
     win = models.BooleanField(verbose_name='Ставка выиграла', default=False)
     date = models.DateTimeField(verbose_name='Дата совершения ставки', auto_now_add=True)
     round_number = models.PositiveBigIntegerField(verbose_name='Номер раунда', default=0)
